requirements.txt
streamlit
cycler==0.10.0
et-xmlfile==1.0.1
jdcal==1.4.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
openpyxl==3.0.6
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.2
xlrd==2.0.1